fix(router): deploy router services for private ARO and remove KAS/OAPI dependencies
Move UseHCPRouter() to a shared util package to break the import cycle
and allow infra.go to use it directly for determining whether router
services should be deployed. This fixes an issue where router services
were not deployed alongside the router deployment for private ARO
clusters (Swift enabled).

Additionally, remove the dependency of the router deployment on KAS
and OAPI components, and add "router" to the list of components that
are allowed to report not-found conditions without blocking overall
status.

Commit-Message-Assisted-by: Claude (via Claude Code)
