+++ control-plane-operator/controllers/hostedcontrolplane/v2/router/util/util.go
@@ -1,0 +1,30 @@
+package util
+
+import (
+	hyperv1 "github.com/openshift/hypershift/api/hypershift/v1beta1"
+	"github.com/openshift/hypershift/support/azureutil"
+	"github.com/openshift/hypershift/support/util"
+)
+
+// UseHCPRouter returns true when the HCP routes should be served by a dedicated
+// HCP router. This occurs when:
+//  1. The cluster is private (e.g. AWS/GCP Private or PublicAndPrivate endpoint access,
+//     or ARO with Swift enabled), OR
+//  2. The cluster is public but uses a dedicated Route for KAS DNS (rather than a LoadBalancer)
+//
+// Excludes IBM Cloud platform.
+func UseHCPRouter(hcp *hyperv1.HostedControlPlane) bool {
+	if hcp.Spec.Platform.Type == hyperv1.IBMCloudPlatform {
+		return false
+	}
+	// For ARO HCP, the dedicated HCP router is only needed when the cluster is private
+	// (Swift enabled). LabelHCPRoutes returns true for all ARO to support the
+	// SharedIngressReconciler, but that doesn't mean a dedicated router deployment is needed.
+	if azureutil.IsAroHCP() {
+		return util.IsPrivateHCP(hcp)
+	}
+	// Router infrastructure is needed when:
+	// 1. Cluster has private access (Private or PublicAndPrivate) - for internal routes, OR
+	// 2. External routes are labeled for HCP router (Public with KAS DNS)
+	return util.IsPrivateHCP(hcp) || util.LabelHCPRoutes(hcp)
+}
